requirements.txt
Wand==0.5.6
Werkzeug==1.0.1
certifi==2020.12.5
click==7.1.2
cycler==0.10.0
decorator==4.4.2
deskew==0.10.28
Flask==1.1.2
imageio==2.9.0
imutils==0.5.4
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
networkx==2.5.1
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.2.0
pyparsing==2.4.7
pytesseract==0.3.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scipy==1.5.3
six==1.13.0
tifffile==2020.9.3
